refactor: Introduce global exception handler using @RestControllerAdvice
- Centralized exception handling for all controllers to reduce code duplication and ensure consistent error responses.
- Removed individual exception handlers from UserController as they are now globally handled.
