+++ src/main/kotlin/user/controller/UserController.kt
@@ -1,7 +1,6 @@
 package com.wafflestudio.seminar.spring2023.user.controller
 
 import com.wafflestudio.seminar.spring2023.user.service.*
-import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
@@ -13,7 +12,7 @@ class UserController(
     fun signUp(
         @RequestBody request: SignUpRequest,
     ): ResponseEntity<Unit> {
-        val user = userService.signUp(
+        userService.signUp(
             request.username,
             request.password,
             request.image,
@@ -49,17 +48,6 @@ class UserController(
             )
         )
     }
-    @ExceptionHandler(SignUpBadUsernameException::class, SignUpBadPasswordException::class)
-    fun handleSignUpBadRequest(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.BAD_REQUEST).build()
-
-    @ExceptionHandler(SignUpUsernameConflictException::class)
-    fun handleSignUpConflict(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.CONFLICT).build()
-
-    @ExceptionHandler(SignInUserNotFoundException::class, SignInInvalidPasswordException::class)
-    fun handleSignInNotFound(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.NOT_FOUND).build()
-
-    @ExceptionHandler(AuthenticateException::class)
-    fun handleAuthenticationFailed(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
 }
 
 data class UserMeResponse(
