+++ src/main/kotlin/user/controller/UserControllerV2.kt
@@ -1,9 +1,6 @@
 package com.wafflestudio.seminar.spring2023.user.controller
 
 import com.wafflestudio.seminar.spring2023.user.service.*
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -38,22 +35,4 @@ class UserControllerV2(
             image = user.image
         )
     }
-
-    @ExceptionHandler(SignUpBadUsernameException::class)
-    fun handleSignUpBadUsername(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.BAD_REQUEST).build()
-
-    @ExceptionHandler(SignUpBadPasswordException::class)
-    fun handleSignUpBadPassword(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.BAD_REQUEST).build()
-
-    @ExceptionHandler(SignUpUsernameConflictException::class)
-    fun handleSignUpConflict(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.CONFLICT).build()
-
-    @ExceptionHandler(SignInUserNotFoundException::class)
-    fun handleSignInUserNotFound(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.NOT_FOUND).build()
-
-    @ExceptionHandler(SignInInvalidPasswordException::class)
-    fun handleSignInInvalidPassword(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.NOT_FOUND).build()
-
-    @ExceptionHandler(AuthenticateException::class)
-    fun handleAuthenticationFailed(): ResponseEntity<Unit> = ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
 }
